% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{del_extra_car}
\alias{del_extra_car}
\title{Function to eliminate extra characters}
\usage{
del_extra_car(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without extra characters
}
\description{
This function eliminates extra characters. For now it only eliminates some common extra characters
}
\examples{
del_extra_car(cjData$conjoint.0.options)
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{characters}
\keyword{clean}
