% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divide_by_x}
\alias{divide_by_x}
\title{Divide by x}
\usage{
divide_by_x(vec, x)
}
\arguments{
\item{vec}{vector to divide by x}

\item{x}{value to divide the vector by}
}
\value{
vector divided by x
}
\description{
This function divides any vector by a given X, we use it regularly to include in pipes.
}
\examples{
divide_by_x(data$percentage, 100) # This creates proportion from percentage
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{divide}
\keyword{hundreds}
\keyword{percentage}
\keyword{proportion}
