% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_0_to_100}
\alias{from_0_to_100}
\title{Function to rescale a vector to 0-100 scale}
\usage{
from_0_to_100(x)
}
\arguments{
\item{x}{numeric vector to rescale to 0-100}
}
\value{
numeric vector in a 0-100 scale
}
\description{
This function takes any vector and rescales it in a linear way so that the new range is 0 to 100. Useful specially when creating indicators
}
\examples{
x <- c(1,2,3,1,3,2,4,5,4,3,2,5,4,3,2,1,4,5)
from_0_to_100(x)
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{indicators}
\keyword{rescale}
