% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaborioRtools-package.R
\docType{package}
\name{gaborioRtools}
\alias{gaborioRtools}
\title{gaborioRTools: My own personal R package}
\source{
\url{https://github.com/SensataUx/sensataDataProg}
}
\description{
This package has some functions that I have created and dependencies that I regularly use to manipulate and clean data.
}
\section{find which data management functions to keep}{

I need to check which function I will keep:
- divide_by_x:  Divide by X
}

\section{find which data analysis functions to keep}{

}

