% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_all_na}
\alias{not_all_na}
\title{Selecter for columns that are ALL NA}
\usage{
not_all_na(x)
}
\arguments{
\item{df}{}
}
\value{
selects all variables where ALL rows are NA
}
\description{
This a selection helper is used normally inside a select call to drop variables that are all NA
}
\examples{
select(where(not_all_na))
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{na}
\keyword{selecter}
