% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-class-doc.r
\name{open_class_doc}
\alias{open_class_doc}
\title{Open Class Document}
\usage{
open_class_doc(class_number, render = FALSE)
}
\arguments{
\item{class_number}{Number of the class to open (e.g., 5 for clase5.Rmd)}

\item{render}{Whether to re-render the document before opening. Default is FALSE}
}
\value{
Opens the HTML document in a web browser
}
\description{
This function renders and opens an R Markdown document for a specific class
}
\examples{
open_class_doc(5) # Opens clase5.html
open_class_doc(6, render = FALSE) # Opens without re-rendering
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{class}
\keyword{documentation}
\keyword{teaching}
