% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_100.R
\name{recode_100}
\alias{recode_100}
\title{Recode to 100 function}
\usage{
recode_100(df, col, values)
}
\arguments{
\item{df}{}
}
\value{
selects all variables where ALL rows are NA
}
\description{
This function recodes a column to 0-100 creating a new one with _r suffix
}
\examples{
cols_values <- list(
 EMO_11 = c("5", "6. Positiva"),
 PAF_11 = c("5", "6. Honestos"),
 PAF_14 = c("5", "6. Solidarios"),
 PAF_03 = c("4", "5. De acuerdo"),
 PAF_04 = c("4", "5. De acuerdo"),
 PAF_08 = c("Smiling", "Very happy")
)
for (col in names(cols_values)) {
    pol_data <- recode_100(pol_data, col, cols_values[[col]])
}
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{na}
\keyword{selecter}
