% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols.R
\name{select_cols}
\alias{select_cols}
\title{Function to export sensata data for final customer}
\usage{
select_cols(
  df,
  dropGeo = T,
  geoCoordinates = F,
  dropParams = T,
  dropUserData = T,
  dropMetaData = T,
  dropTotalTime = T,
  dropQuestionTime = T
)
}
\arguments{
\item{df}{data downloaded from Mongo, cleaned with cleanData.R, scrubbed with scrubData.R and prepared with factorSensata.R.}

\item{dropGeo}{if TRUE will drop coordinates (lat & long) and geo.accuracy}

\item{geoCoordinates}{if TRUE will drop geo.coordinates columns, instead of lat & long}

\item{dropParams}{if TRUE will drop all columns that start with params}

\item{dropUserData}{if TRUE will drop fingerprint and sensataId}

\item{dropMetaData}{if TRUE will drop createdAt, surveyId and surveyName}

\item{dropTotalTime}{if TRUE it drops totalTimeMin column}

\item{dropQuestionTime}{if TRUE drops q_time columns}
}
\value{
Dataframe ready for client.
}
\description{
This function eliminates most metadata from the dataset. This is basically a wrapper for tidyverse::select()
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gabriel.n.c.t182@gmail.com}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
